requirements.txt
flask==2.0.1
werkzeug==2.0.1
flask-cors==3.0.10
flask-socketio==5.1.1
eventlet==0.33.0
python-engineio==4.2.1
python-socketio==5.4.0
numpy
Pillow
# For real hardware implementation
RPi.GPIO==0.7.1
picamera2
libcamera
# pyaudio==0.2.12
# whisper==1.0.0
# ollama==0.1.0